Change LoginMenu to MainMenu
