--- src/scivu/scivu/Views/LoginMenuView.axaml.cs
@@ -1,13 +1,0 @@
-﻿using Avalonia;
-using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
-
-namespace scivu.Views;
-
-public partial class LoginMenuView : UserControl
-{
-    public LoginMenuView()
-    {
-        InitializeComponent();
-    }
-}
